+++ tasty-plutus/CHANGELOG.md
@@ -4,6 +4,17 @@ This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 
 ## Unreleased
 
+## 7.2 - 2022-01-30
+
+### Added
+
+* Record fields for `ContextBuilder`:
+  * `cbInputs`
+  * `cbOutputs`
+  * `cbSignatories`
+  * `cbDatums`
+  * `cbMinting`
+
 ## 7.1 -- 2022-01-30
 
 ### Added
@@ -15,8 +26,8 @@ This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 ## 7.0 -- 2022-01-21
 
 ### Changed
-
-* Plutus upgrade: `plutus` pinned to `65bad0fd53e432974c3c203b1b1999161b6c2dce`,
+  
+* Plutus upgrade: `plutus` pinned to `65bad0fd53e432974c3c203b1b1999161b6c2dce`, 
   `plutus-apps` pinned to `34fe6eeff441166fee0cd0ceba68c1439f0e93d2`
 
 ## 6.0 -- 2022-01-19